kvm: fix net long names
Update kvm flavor networking code with bits added in CNI.

Resolves #2533
Reverts #2538
